+++ src/components/ProductCard/ProductCard.tsx
@@ -2,7 +2,7 @@ import { memo } from "react";
 import { ReactComponent as ShoppingCartImg } from "../../assets/icon/shopping-cart.svg";
 import Counter from "../common/Counter/Counter";
 import ProductImg from "../common/ProductImg/ProductImg";
-import ProductInfo from "../ProductInfo/ProductInfo";
+import ProductInfo from "./ProductInfo/ProductInfo";
 import useCart from "../../hooks/useCart";
 import Styled from "./ProductCardStyled";
