+++ src/components/ProductCard/ProductInfo/ProductInfoStyled.ts
